# src/main/java/de/muenchen/dave/geodataeai/api/dto/error/InformationResponseDto.java
/*
 * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
 * der Landeshauptstadt München, 2023
 */
package de.muenchen.dave.geodataeai.api.dto.error;

import de.muenchen.dave.geodataeai.api.dto.enums.InformationResponseType;
import java.time.LocalDateTime;
import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class InformationResponseDto {

    private InformationResponseType type;

    private LocalDateTime timestamp;

    private String traceId;

    private String spanId;

    private Integer httpStatus;

    private String originalException;

    private List<String> messages;
}