# src/main/java/de/muenchen/dave/geodataeai/api/dto/messwerte/TagesaggregatDto.java
/*
 * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
 * der Landeshauptstadt München, 2023
 */
package de.muenchen.dave.geodataeai.api.dto.messwerte;

import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.ToString;

@Data
@ToString(callSuper = true)
@EqualsAndHashCode(callSuper = true)
public class TagesaggregatDto extends MesswertDto {
    // Messtage

    private Long includedMeasuringDays;
}